Rename loadPage to visit
